Fix urxvtd crash when using a background expression.
A rxvt_img object contains a pointer to the associated rxvt_screen,
which in practice corresponds to the terminal object that created the
image. This causes a use-after-free when there are two or more terminal
objects sharing a rxvt_img and the one that created it is destroyed.
This patch fixes the issue by replacing the rxvt_screen pointer with a
pointer to the associated rxvt_display.